refactor(db): move partition maintenance store ownership
