+++ crates/buzz-db/src/lib.rs
@@ -1138,33 +1138,6 @@ impl Db {
             }
         }
     }
-
-    /// Ensures monthly partitions exist for the next N months.
-    #[datastore_span(name = "ensure_future_partitions", system = "postgresql")]
-    pub async fn ensure_future_partitions(&self, months_ahead: u32) -> Result<()> {
-        partition::ensure_future_partitions(&self.pool, months_ahead).await
-    }
-
-    /// Backfill `d_tag` for existing NIP-33 events (kind 30000–39999) that have `d_tag IS NULL`.
-    ///
-    /// Idempotent — safe to call on every startup. No-ops when all rows are already populated.
-    /// Runs a single UPDATE touching only NIP-33 rows with NULL d_tag.
-    #[datastore_span(name = "backfill_d_tags", system = "postgresql")]
-    pub async fn backfill_d_tags(&self) -> Result<u64> {
-        let result = sqlx::query(
-            "UPDATE events \
-             SET d_tag = COALESCE( \
-                 (SELECT elem->>1 FROM jsonb_array_elements(tags) AS elem \
-                  WHERE elem->>0 = 'd' LIMIT 1), \
-                 '' \
-             ) \
-             WHERE kind BETWEEN 30000 AND 39999 AND d_tag IS NULL \
-               AND community_write_allowed(community_id)",
-        )
-        .execute(&self.pool)
-        .await?;
-        Ok(result.rows_affected())
-    }
 }
 
 #[cfg(test)]
